requirements.txt
pygame==2.6.1
python-vlc==3.0.21203